Support new admin-bro ValidationError parameters
Version 2.2.0 of `admin-bro` changed the parameters to `ValidationError`
so that there is no generic `message` as the first argument. This was
done in [this
commit](https://github.com/SoftwareBrothers/admin-bro/commit/6b937c3077a100fa9c16cdf3d9661b36c6c48277).

So, removed the first argument and upgraded the `admin-bro` peer
dependency to `>=2.2.0`.